Split loader header per type
