+++ src/runtime/loader_v1_any.h
@@ -1,0 +1,58 @@
+#ifndef SOURCEMETA_JSONBINPACK_LOADER_V1_ANY_H_
+#define SOURCEMETA_JSONBINPACK_LOADER_V1_ANY_H_
+
+#include <sourcemeta/jsonbinpack/runtime.h>
+
+#include <sourcemeta/jsontoolkit/json.h>
+
+#include <cassert> // assert
+#include <utility> // std::move
+#include <vector>  // std::vector
+
+namespace sourcemeta::jsonbinpack::v1 {
+
+auto BYTE_CHOICE_INDEX(const sourcemeta::jsontoolkit::JSON &options) -> Plan {
+  assert(options.defines("choices"));
+  const auto &choices{options.at("choices")};
+  assert(choices.is_array());
+  const auto &array{choices.as_array()};
+  std::vector<sourcemeta::jsontoolkit::JSON> elements{array.cbegin(),
+                                                      array.cend()};
+  return sourcemeta::jsonbinpack::BYTE_CHOICE_INDEX({std::move(elements)});
+}
+
+auto LARGE_CHOICE_INDEX(const sourcemeta::jsontoolkit::JSON &options) -> Plan {
+  assert(options.defines("choices"));
+  const auto &choices{options.at("choices")};
+  assert(choices.is_array());
+  const auto &array{choices.as_array()};
+  std::vector<sourcemeta::jsontoolkit::JSON> elements{array.cbegin(),
+                                                      array.cend()};
+  return sourcemeta::jsonbinpack::LARGE_CHOICE_INDEX({std::move(elements)});
+}
+
+auto TOP_LEVEL_BYTE_CHOICE_INDEX(const sourcemeta::jsontoolkit::JSON &options)
+    -> Plan {
+  assert(options.defines("choices"));
+  const auto &choices{options.at("choices")};
+  assert(choices.is_array());
+  const auto &array{choices.as_array()};
+  std::vector<sourcemeta::jsontoolkit::JSON> elements{array.cbegin(),
+                                                      array.cend()};
+  return sourcemeta::jsonbinpack::TOP_LEVEL_BYTE_CHOICE_INDEX(
+      {std::move(elements)});
+}
+
+auto CONST_NONE(const sourcemeta::jsontoolkit::JSON &options) -> Plan {
+  assert(options.defines("value"));
+  return sourcemeta::jsonbinpack::CONST_NONE({options.at("value")});
+}
+
+auto ANY_PACKED_TYPE_TAG_BYTE_PREFIX(const sourcemeta::jsontoolkit::JSON &)
+    -> Plan {
+  return sourcemeta::jsonbinpack::ANY_PACKED_TYPE_TAG_BYTE_PREFIX{};
+}
+
+} // namespace sourcemeta::jsonbinpack::v1
+
+#endif
